validation added to check that the argument passed is a file

+++ file_deletion.py
@@ -3,5 +3,5 @@ import os
 
 def file_delete(filename):
     """deletes files from the filesystem"""
-    os.remove(filename)
-    # print(f"File {filename} deleted successfully")
+    if os.path.isfile(filename):
+        os.remove(filename)
